Move todolists to own directory
